- 배송조회API 웹훅연동 재처리 - 배송조회API 이벤트 발행으로 호출 - Req, Resp 클래스 이름 통일 - TopItemSales 테이블 추가

+++ BackEnd/pcgear/src/main/java/com/pcgear/complink/pcgear/Dashboard/TopItemSalesDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class TopItemSalesDto {
     private Integer itemId;
     private String itemName;
-    private Long quantitySold;
+    private Integer quantitySold;
     private Long totalRevenue;
-    private Long currentStock;
+    private Integer currentStock;
 }
